pay-sdk-py/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voidly-pay"
version = "1.0.1"
description = "Python SDK for Voidly Pay — the off-chain credit ledger + hire marketplace for AI agents. Lets any Python agent bootstrap, pay, hire, and settle with other agents via Ed25519-signed envelopes. Ships with LangChain / CrewAI / AutoGen integration examples."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "Voidly Research", email = "research@voidly.ai" }]
keywords = [
  "voidly",
  "voidly-pay",
  "ai-agents",
  "agent-payments",
  "marketplace",
  "escrow",
  "ed25519",
  "micro-payments",
  "autonomous-agents",
  "crewai",
  "langchain",
  "autogen",
  "mcp",
]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Office/Business :: Financial",
]
dependencies = ["requests>=2.28", "pynacl>=1.5"]

[project.urls]
Repository = "https://github.com/voidly-ai/voidly-pay/tree/main/pay-sdk-py"
Homepage = "https://voidly.ai/pay"
Documentation = "https://voidly.ai/voidly-pay-for-ai-agents.md"
Source = "https://github.com/voidly-ai/voidly-pay/tree/main/pay-sdk-py"
Issues = "https://github.com/voidly-ai/voidly-pay/issues"

[tool.setuptools.packages.find]
where = ["src"]
